split header parsing out into separate module for use by rest_json and rest_xml fix bugs in type filter turn on deny(dead_code)
